requirements.txt
fastapi
uvicorn
feedparser
requests
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-trace
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-propagator-gcp
